+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Workbench as requirement for Clutter pieces missing a workstation requirement.
 
+### Fixed
+
+- Fix MonsterLabZ spiders taking fall damage.
+
 ## [0.3.0] - 2024-05-04
 
 ### Changed